Add check for min_gap to be atleast 4 hours in rain_events
According to the DWA A-531 Chapter 4.2, rain events should be atleast 4 hours apart to be statistically indepedent. This period is defined as the default parameter in the rain_events function. However, during function calls in e.g. calculate_u_uw() smaller gap_periods are used. Therefore a new check is added, such that even if the method is called with a min_gap < 4 hours, it still will be atleast 4 hours.